+++ tests/ops_data_store_tests/cli/test_cli_config.py
@@ -1,9 +1,12 @@
 from io import StringIO
 from pprint import pprint
 
+import pytest
+from pytest_mock import MockerFixture
 from typer.testing import CliRunner
 
 from ops_data_store.cli import app as cli
+from ops_data_store.config import Config
 
 
 class TestCliConfigShow:
@@ -20,3 +23,19 @@ class TestCliConfigShow:
 
         assert result.exit_code == 0
         assert expected in result.output
+
+    def test_invalid_config(
+        self, caplog: pytest.LogCaptureFixture, fx_cli_runner: CliRunner, mocker: MockerFixture
+    ) -> None:
+        """Returns error message where config is invalid."""
+        mocker.patch.object(
+            target=Config, attribute="validate", side_effect=RuntimeError("Required config option `DB_DSN` not set.")
+        )
+
+        result = fx_cli_runner.invoke(app=cli, args=["config", "show"])
+
+        assert "Checking app config." in caplog.text
+        assert "Required config option `DB_DSN` not set." in caplog.text
+
+        assert result.exit_code == 1
+        assert "No. Application config is invalid. Check with `ods-ctl config check`." in result.output